fix: exit 버튼에 의한 fullscreen 컨트롤 - exit으로 풀스크린 취소
